feat: Job 支持多租户 #3369
完善初始化租户逻辑，增加初始化租户内消息通知渠道
