+++ support-files/kubernetes/charts/bk-job/templates/job-assemble/configmap.yaml
@@ -344,7 +344,7 @@ data:
         notify:
           default:
             channels:
-              available: mail
+              available: mail,weixin,rtx
       execute:
         # 执行日志导出产生的临时文件相关配置        
         log-export:
