+++ support-files/kubernetes/charts/bk-job/templates/job-manage/configmap.yaml
@@ -126,7 +126,7 @@ data:
         notify:
           default:
             channels:
-              available: mail
+              available: mail,weixin,rtx
     mysql:
       {{- toYaml .Values.manageConfig.mysql | nindent 6 }}
     server:
