+++ support-files/templates/#etc#job#job-manage#job-manage.yml
@@ -90,6 +90,6 @@ job:
     notify:
       default:
         channels:
-          available: mail
+          available: mail,weixin,rtx
